+++ packages/flutter_meedu/lib/src/state/widgets/watch_filter.dart
@@ -4,7 +4,7 @@ import 'package:meedu/provider.dart';
 import 'package:meedu/state.dart';
 
 typedef _BuildWhen<S> = bool Function(S prev, S current);
-typedef _BuildBySelect<T, S> = S Function(T);
+typedef _BuildBySelect<Notifier, Result> = Result Function(Notifier);
 
 /// enum used to identifier filters used in a provider
 enum Filter {
@@ -14,14 +14,10 @@ enum Filter {
   /// .when filter fot StateProvider
   /// this filter only can be  used with ref.watch
   when,
-
-  /// .ids filter for SimpleProvider
-  /// this filter only can be  used with ref.watch
-  ids,
 }
 
 /// class to save a Notifier, the listener and the rebuild function
-class Target<Notifier, R> extends Provider<Notifier> {
+class Target<Notifier, Result> extends Provider<Notifier> {
   // ignore: public_member_api_docs
   Target(this.notifier);
 
@@ -35,10 +31,10 @@ class Target<Notifier, R> extends Provider<Notifier> {
   void Function()? rebuild;
 
   /// used to store the value returned by .select or .when
-  late R selectValue;
+  late Result selectValue;
 
   /// callback defined when a filter is used
-  dynamic callback;
+  late Function callback;
 
   /// filter type
   late Filter filter;
@@ -48,52 +44,12 @@ class Target<Notifier, R> extends Provider<Notifier> {
 extension SimpleProviderExt<Notifier> on SimpleProvider<Notifier> {
   /// use this method to rebuild your Consumer when a value has changed
   /// or you can use a boolean condition. Check the documentation for more info.
-  Target<Notifier, R> select<R>(_BuildBySelect<Notifier, R> cb) {
+  Target<Notifier, Result> select<Result>(
+      _BuildBySelect<Notifier, Result> callback) {
     // get the  Notifier attached to this SimpleProvider
-    final target = Target<Notifier, R>(read);
+    final target = Target<Notifier, Result>(read);
     target.filter = Filter.select;
-    target.callback = cb;
-    return target;
-  }
-
-  /// use this method to rebuild your Consumer using ids (a list of strings)
-  /// passed when you call [notify(['id1','id2',...])]
-  ///
-  /// If you pass to notify method using
-  /// an empty list of ids you can use [allowNotifyWithEmptyIds]
-  /// to decide if the Consumer widget or a ProviderListener must be notified
-  @Deprecated('the .ids filter is obsolete in favor to .select'
-      ' and it will be removed in flutter_meedu:^6.x.x')
-  Target<Notifier, List> ids(
-    List<String> Function() cb, {
-    bool allowNotifyWithEmptyIds = true,
-  }) {
-    final target = Target<Notifier, List>(read);
-    target.filter = Filter.ids;
-    // ignore: prefer_function_declarations_over_variables
-    final listener = (_) {
-      final List<String> listeners = _;
-      // get the ids passed in the notify method
-      // get the ids returned in the callback
-      final ids = cb();
-      if (listeners.isNotEmpty) {
-        // if the current Consumer contains at least one id inside the listeners
-        for (final id in ids) {
-          if (listeners.contains(id)) {
-            if (target.rebuild != null) {
-              target.rebuild!();
-            }
-            break;
-          }
-        }
-      } else if (allowNotifyWithEmptyIds) {
-        // update the widget if listeners is empty
-        if (target.rebuild != null) {
-          target.rebuild!();
-        }
-      }
-    };
-    target.listener = listener as dynamic;
+    target.callback = callback;
     return target;
   }
 }
@@ -103,35 +59,34 @@ extension StateProviderExt<Notifier extends StateNotifier<S>, S>
     on StateProvider<Notifier, S> {
   /// use this method to rebuild your Consumer using the previous state and the current
   /// state to return a boolean
-  Target<Notifier, bool> when(_BuildWhen<S> cb) {
+  Target<Notifier, bool> when(_BuildWhen<S> callback) {
     final target = Target<Notifier, bool>(read);
     target.filter = Filter.when;
-    target.callback = cb;
+    target.callback = callback;
     return target;
   }
 
   /// use this method to rebuild your Consumer when a value in the state has changed
   /// or you can use a boolean condition. Check the documentation for more info.
-  Target<Notifier, R> select<R>(_BuildBySelect<S, R> cb) {
-    final target = Target<Notifier, R>(read);
+  Target<Notifier, Result> select<Result>(_BuildBySelect<S, Result> callback) {
+    final target = Target<Notifier, Result>(read);
     target.filter = Filter.select;
-    target.callback = cb;
+    target.callback = callback;
     return target;
   }
 }
 
 /// create the listener for provider.select filter (SimpleProvider)
 void createSimpleSelectListener(Target target) {
-  final cb = target.callback;
   final notifier = target.notifier;
   // get an initial value using the callback
-  dynamic prevValue = cb(notifier);
+  dynamic prevValue = target.callback(notifier);
   target.selectValue = prevValue;
 
   // listener with  the logic to rebuild the Consumer
   // ignore: prefer_function_declarations_over_variables
   final listener = (_) {
-    final value = cb(notifier);
+    final value = target.callback(notifier);
     target.selectValue = value;
     // check if the value has changed
     if (prevValue != value || (value is bool && value)) {
@@ -147,14 +102,12 @@ void createSimpleSelectListener(Target target) {
 
 /// create the listener for provider.select filter (StateProvider)
 void createStateSelectListener(Target target) {
-  // get an initial value using the callback
-  final cb = target.callback;
-  dynamic prevValue = cb(target.notifier.state);
+  dynamic prevValue = target.callback(target.notifier.state);
   target.selectValue = prevValue;
 
   // ignore: prefer_function_declarations_over_variables
   final listener = (newState) {
-    final value = cb(target.notifier.state);
+    final value = target.callback(target.notifier.state);
     target.selectValue = value;
     // check if the value has changed
     if (prevValue != value || (value is bool && value)) {
@@ -169,14 +122,13 @@ void createStateSelectListener(Target target) {
 
 /// create the listener for provider.when filter
 void createWhenListener(Target target) {
-  final cb = target.callback;
   final notifier = target.notifier as StateNotifier;
-  target.selectValue = cb(notifier.state, notifier.state);
+  target.selectValue = target.callback(notifier.state, notifier.state);
 
   // ignore: prefer_function_declarations_over_variables
   final listener = (newState) {
     // rebuild the Consumer using the boolean returned by the callback
-    final allowRebuild = cb(notifier.oldState, newState);
+    final allowRebuild = target.callback(notifier.oldState, newState);
     target.selectValue = allowRebuild;
     if (allowRebuild) {
       if (target.rebuild != null) {