+++ packages/flutter_meedu/test/listeners/provider_listener_test_with_filters_test.dart
@@ -49,64 +49,6 @@ void main() {
       }
     },
   );
-
-  testWidgets(
-    'ProviderListener > SimpleProvider filters > ids',
-    (tester) async {
-      int numberOfCalls = 0;
-      int minValueToBeNotified = 5;
-      await tester.pumpWidget(
-        MaterialApp(
-          home: ProviderListener<_CounterController>(
-            provider: _counterProvider.ids(
-              () => ['id'],
-              allowNotifyWithEmptyIds: false,
-            ),
-            onChange: (_, __) {
-              numberOfCalls++;
-            },
-            builder: (_, controller) {
-              return Scaffold(
-                body: Center(
-                  child: Column(
-                    children: [
-                      TextButton(
-                        child: Text("add"),
-                        onPressed: controller.incremment,
-                      ),
-                      TextButton(
-                        child: Text("addId1"),
-                        onPressed: controller.incremmentWithId1,
-                      ),
-                      TextButton(
-                        child: Text("addId2"),
-                        onPressed: controller.incremmentWithId2,
-                      ),
-                    ],
-                  ),
-                ),
-              );
-            },
-          ),
-        ),
-      );
-      final controller = _counterProvider.read;
-      for (int i = 1; i <= minValueToBeNotified + 3; i++) {
-        await tester.tap(find.text("addId2"));
-        await tester.pump();
-        expect(controller.counter, i);
-        expect(numberOfCalls, 0);
-      }
-      await tester.tap(find.text("addId1"));
-      await tester.pump();
-      await tester.tap(find.text("addId1"));
-      await tester.pump();
-      expect(numberOfCalls, 2);
-      await tester.tap(find.text("add"));
-      await tester.pump();
-      expect(numberOfCalls, 2);
-    },
-  );
 }
 
 final _counterProvider = SimpleProvider(
@@ -120,15 +62,5 @@ class _CounterController extends SimpleNotifier {
   void incremment() {
     _counter++;
     notify();
-  }
-
-  void incremmentWithId1() {
-    _counter++;
-    notify(['id']);
-  }
-
-  void incremmentWithId2() {
-    _counter++;
-    notify(['id2']);
   }
 }