+++ packages/meedu/lib/src/notifiers/simple_notifier.dart
@@ -2,20 +2,10 @@ import 'package:meta/meta.dart' show mustCallSuper;
 
 import 'base_notifier.dart';
 
-abstract class SimpleNotifier extends BaseNotifier<List<String>> {
+abstract class SimpleNotifier extends BaseNotifier {
   /// notify to listeners and rebuild the widgets
-  ///
-  /// [listeners] a list of strings to update the widgets (MeeduBuilder) with the ids inside the list
-  @override
-  void notify([
-    @Deprecated(
-      'This property is obsolete because it was used for `.ids`'
-      ' filter but `.select` filter more efficient so this property will be removed'
-      ' in meedu:^5.x.x and flutter_meedu:^6.x.x',
-    )
-        List<String> listeners = const [],
-  ]) {
-    super.notify(listeners);
+  void notify() {
+    notifyListeners([]);
   }
 
   /// use to listen when the controller was deleted from memory